+++ security/test_signing.py
@@ -1,16 +1,20 @@
 import hashlib
 import json
+from datetime import datetime, timedelta
 from unittest.mock import patch
 
 import http_sfv
 import pytest
 import requests
 from cryptography import x509
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.serialization import (
     load_pem_private_key,
     load_pem_public_key,
 )
+from cryptography.x509.oid import NameOID
 from django.test import TestCase
 from http_message_signatures import (
     HTTPMessageSigner,
@@ -22,11 +26,81 @@ from http_message_signatures import (
 from jwcrypto import jwk, jws
 
 import security.helper as helper
-from security.signing import (
-    BlenderCertificationResolver,
-    MessageVerifier,
-    ResponseSigner,
-)
+from security.signing import MessageVerifier, ResponseSigner
+
+
+class _TestCertificationCreator:
+    def create_csr(self, pvt_key):
+        csr = (
+            x509.CertificateSigningRequestBuilder()
+            .subject_name(
+                x509.Name(
+                    [
+                        x509.NameAttribute(NameOID.COUNTRY_NAME, "FI"),
+                        x509.NameAttribute(
+                            NameOID.STATE_OR_PROVINCE_NAME, "TEST_Pirkanmaa"
+                        ),
+                        x509.NameAttribute(NameOID.LOCALITY_NAME, "TEST_Tampere"),
+                        x509.NameAttribute(NameOID.ORGANIZATION_NAME, "TEST_TII"),
+                        x509.NameAttribute(NameOID.COMMON_NAME, "test.blender.api"),
+                    ]
+                )
+            )
+            .add_extension(
+                x509.SubjectAlternativeName(
+                    [
+                        # Describe what sites we want this certificate for.
+                        x509.DNSName("http://localhost:8000/"),
+                    ]
+                ),
+                critical=False,
+                # Sign the CSR with private key.
+            )
+            .sign(pvt_key, hashes.SHA256())
+        )
+        return csr
+
+    def create_ca(self, pvt_key, csr):
+        # Create a CA certificate
+        ca_subject = x509.Name(
+            [
+                x509.NameAttribute(x509.NameOID.COUNTRY_NAME, "FI"),
+                x509.NameAttribute(
+                    x509.NameOID.STATE_OR_PROVINCE_NAME, "TEST_Pirkanmaa"
+                ),
+                x509.NameAttribute(x509.NameOID.LOCALITY_NAME, "TEST_Tampere"),
+                x509.NameAttribute(x509.NameOID.ORGANIZATION_NAME, "TEST_TII"),
+                x509.NameAttribute(x509.NameOID.COMMON_NAME, "test.blender.api"),
+            ]
+        )
+        not_valid_before_time = datetime.utcnow()
+        not_valid_after_time = not_valid_before_time + timedelta(days=365)
+        ca_certificate = (
+            x509.CertificateBuilder()
+            .subject_name(ca_subject)
+            .issuer_name(ca_subject)
+            .public_key(pvt_key.public_key())
+            .serial_number(x509.random_serial_number())
+            .add_extension(
+                x509.BasicConstraints(ca=True, path_length=None), critical=True
+            )
+            .not_valid_before(not_valid_before_time)
+            .not_valid_after(not_valid_after_time)
+            .sign(pvt_key, hashes.SHA256(), default_backend())
+        )
+
+        # Issue a certificate using the CSR and CA
+        issued_certificate = (
+            x509.CertificateBuilder()
+            .subject_name(csr.subject)
+            .issuer_name(ca_certificate.subject)
+            .public_key(csr.public_key())
+            .serial_number(x509.random_serial_number())
+            .not_valid_before(not_valid_before_time)
+            .not_valid_after(not_valid_after_time)
+            .sign(pvt_key, hashes.SHA256(), default_backend())
+        )
+        return issued_certificate
 
 
 class _TestHTTPSignatureKeyResolver(HTTPSignatureKeyResolver):
@@ -40,16 +114,16 @@ class _TestHTTPSignatureKeyResolver(HTTPSignatureKeyResolver):
 
 
 class _TestCertificationResolver:
-    def __init__(self, pvt_key, cert_resolver: BlenderCertificationResolver):
+    def __init__(self, pvt_key, cert_creator: _TestCertificationCreator):
         self._pvt_key = pvt_key
-        self._cert_resolver = cert_resolver
+        self.cert_creator = cert_creator
 
     def resolve_private_key(self, key_id: str):
         raise NotImplementedError
 
     def resolve_public_key(self, key_id: str):
-        csr = self._cert_resolver.create_csr(pvt_key=self._pvt_key)
-        ca = self._cert_resolver.create_ca(pvt_key=self._pvt_key, csr=csr)
+        csr = self.cert_creator.create_csr(pvt_key=self._pvt_key)
+        ca = self.cert_creator.create_ca(pvt_key=self._pvt_key, csr=csr)
         return ca.public_key()
 
 
@@ -183,7 +257,7 @@ class CertTests(TestCase):
     def setUp(self):
         key_resolver = _TestHTTPSignatureKeyResolver()
         self.pvt_key = key_resolver.resolve_private_key(key_id="001")
-        self.cert_resolver = BlenderCertificationResolver()
+        self.cert_creator = _TestCertificationCreator()
 
         _payload = {"type_of_operation": 1, "originating_party": "TII"}
         _request = requests.Request(
@@ -194,12 +268,12 @@ class CertTests(TestCase):
         self.request = _request.prepare()
 
     def test_create_csr(self):
-        csr = self.cert_resolver.create_csr(pvt_key=self.pvt_key)
+        csr = self.cert_creator.create_csr(pvt_key=self.pvt_key)
         self.assertIsInstance(csr, x509.CertificateSigningRequest)
 
     def test_create_ca(self):
-        csr = self.cert_resolver.create_csr(pvt_key=self.pvt_key)
-        ca = self.cert_resolver.create_ca(pvt_key=self.pvt_key, csr=csr)
+        csr = self.cert_creator.create_csr(pvt_key=self.pvt_key)
+        ca = self.cert_creator.create_ca(pvt_key=self.pvt_key, csr=csr)
         self.assertIsInstance(ca, x509.Certificate)
         public_key = ca.public_key()
         self.assertIsInstance(public_key, rsa.RSAPublicKey)
@@ -223,7 +297,7 @@ class CertTests(TestCase):
         verifier = HTTPMessageVerifier(
             signature_algorithm=algorithms.RSA_PSS_SHA512,
             key_resolver=_TestCertificationResolver(
-                pvt_key=self.pvt_key, cert_resolver=self.cert_resolver
+                pvt_key=self.pvt_key, cert_creator=self.cert_creator
             ),
         )
         output = verifier.verify(signed_response)
